Move Orbit.Source -> OC.Source
Move `transformed` and `transformLog` from `Source` class to
`Transformable` interface. Now all methods, events, and logs
related to "transforming" sources are contained in one interface.

Furthermore, Orbit itself now has no concrete class representing
a `Source`. However, OC introduces its own `Source` base class and
enforces the need for a schema to be set on it.
