+++ src/orbit/transformable.js
@@ -1,5 +1,6 @@
 import Orbit from './main';
 import { extend } from './lib/objects';
+import Evented from './evented';
 import Transform from './transform';
 import TransformLog from './transform/log';
 
@@ -8,14 +9,14 @@ export default {
    Mixes the `Transformable` interface into a source
 
    @method extend
-   @param {Source} source Source to extend
-   @returns {Source} Extended source
+   @param {Object} source - Source to extend
+   @returns {Object} Extended source
    */
   extend(source) {
     if (source._transformable === undefined) {
-      source.transformLog = new TransformLog();
-
+      Evented.extend(source);
       extend(source, this.interface);
+      source.transformLog = new TransformLog();
     }
     return source;
   },
@@ -32,6 +33,21 @@ export default {
 
       return this._transform(transform)
         .then(result => this.transformed(result));
+    },
+
+    transformed(transforms) {
+      return transforms
+        .reduce((chain, transform) => {
+          return chain.then(() => {
+            if (this.transformLog.contains(transform.id)) {
+              return Orbit.Promise.resolve();
+            }
+
+            this.transformLog.append(transform.id);
+            return this.settle('transform', transform);
+          });
+        }, Orbit.Promise.resolve())
+        .then(() => transforms);
     }
   }
 };
