+++ test/tests/orbit-common/unit/jsonapi-source-test.js
@@ -1,4 +1,4 @@
-import Source from 'orbit/source';
+import Source from 'orbit-common/source';
 import { uuid } from 'orbit/lib/uuid';
 import Schema from 'orbit-common/schema';
 import KeyMap from 'orbit-common/key-map';
